# More varlen optimizations