--- a/conch/ops/attention/varlen_attention.py
+++ b/conch/ops/attention/varlen_attention.py
@@ -140,13 +140,16 @@ def _check_seqlen_size_compatibility(seq_lens: torch.Tensor, batch_size: int) ->
 
 def _determine_max_num_kv_splits(max_seqlen_q: int, max_seqlen_k: int, max_num_blocks_per_sequence: int) -> int:
     # If we have any prefills (max_seqlen_q > 1), disable FlashDecoding/KV-splits.
-    # If we have all decodes, only enable FlashDecoding if we have a long sequence (>=2048 tokens) and
-    # many cache blocks for each sequence (>=16 cache blocks for the longest sequence).
-    # Note: this is a pretty basic heuristic, we could try and do something more sophisticated in the future
-    if max_seqlen_q > 1 and max_seqlen_k >= 2048 and max_num_blocks_per_sequence >= 16:
+    # If we have all decodes, only enable FlashDecoding if we have a long sequence (>=4096 tokens) and
+    # many cache blocks for each sequence (>=64 cache blocks for the longest sequence).
+    # Note: this is a pretty basic heuristic, we could try and do something more sophisticated in the future. It'd
+    # probably be more efficient to tune this per-platform, but we can't just autotune it because it determines how
+    # much scratchpad memory we allocate. FlashAttention's heuristic uses the number of SMs, it might be worth
+    # considering something like that, but preferably platform-agnostic.
+    if max_seqlen_q == 1 and max_seqlen_k >= 4096 and max_num_blocks_per_sequence >= 64:
         # This number of KV splits affects the size of the scratchpad memory that we allocate,
         # so cap the number of splits at 64 so that we don't allocate too much memory
-        return min(max_num_blocks_per_sequence // 4, 64)
+        return min(max_num_blocks_per_sequence // 16, 64)
 
     return 1