--- a/tools/benchmarks/collect_varlen_attention_decode.sh
+++ b/tools/benchmarks/collect_varlen_attention_decode.sh
@@ -25,6 +25,8 @@ sequence_lengths=(
   "8192"
   "16384"
   "32768"
+  "65536"
+  "131072"
 )
 
 for seq_len in ${sequence_lengths[@]}; do